backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic[email]>=2.0.0
pydantic-settings>=2.7.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
redis>=5.2.0
httpx>=0.28.0
openai>=1.68.0
tavily-python>=0.5.0
